# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio blog JSON API",
        "description": "Portfolio blog API documentation",
        "contact": {
            "email": "yseffk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/blog/items": {
            "get": {
                "tags": [
                    "Admin Content"
                ],
                "summary": "Get Content Items  whith pagination",
                "description": "Display a listing of the resource.",
                "operationId": "79d603ec9bd2fd1028541854025e2cd3",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "description": "Need pagination",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "filtering by title ",
                        "required": false,
                        "example": "Lorem.."
                    },
                    {
                        "name": "is_published",
                        "in": "query",
                        "description": "filtering by 'is published' flag",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "is_free",
                        "in": "query",
                        "description": "filtering by 'is free' flag",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "post_id",
                        "in": "query",
                        "description": "filtering by 'Month block item' ID",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Per page limit",
                        "required": false,
                        "example": "15"
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort  list criteria. Allowed fields are: id, sort, created_at. Prepend '-' sign to revert order",
                        "required": false,
                        "example": "-id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Content"
                ],
                "summary": "Create Content item.",
                "operationId": "ad260c43f2505d2b3077d406f53f647c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "title",
                                    "is_published",
                                    "sort"
                                ],
                                "properties": {
                                    "post_id": {
                                        "description": "Type item ID",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "title": {
                                        "description": "Item title",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "is_published": {
                                        "description": "Item published flag",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "sort": {
                                        "description": "sort",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "brief_content": {
                                        "description": "",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "raw_content": {
                                        "description": "",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Validation information.",
                                            "type": "string",
                                            "example": "The given data was invalid.",
                                            "nullable": false
                                        },
                                        "errors": {
                                            "description": "Object with all errors.",
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/blog/item-attachments/{id}": {
            "get": {
                "tags": [
                    "showing"
                ],
                "summary": "Get item-attachment collection by id",
                "operationId": "be93e21daa30ffca585f0bdfa5aaf70c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item-attachment id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/blog/items/{id}": {
            "get": {
                "tags": [
                    "showing"
                ],
                "summary": "Get content item and attachments",
                "operationId": "b3a347cf49a05e4c3f433c842ab6607c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/blog/posts/blog": {
            "get": {
                "tags": [
                    "showing"
                ],
                "summary": "Get blog items published data by slug.  ",
                "operationId": "be191fbbcc2aba33eac5849a47186efd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Per page limit",
                        "required": false,
                        "example": "15"
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort  list criteria. Allowed fields are: id, sort, published_at. Prepend '-' sign to revert order",
                        "required": false,
                        "example": "-id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/blog/posts/gallery": {
            "get": {
                "tags": [
                    "showing"
                ],
                "summary": "Get gallery items published data by slug.  ",
                "operationId": "72ae4218254e140cd9d5f8834cc5fe01",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Per page limit",
                        "required": false,
                        "example": "15"
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort  list criteria. Allowed fields are: id, sort, published_at. Prepend '-' sign to revert order",
                        "required": false,
                        "example": "-id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}